profile page

+++ app/Http/Controllers/IdeaController.php
@@ -21,12 +21,14 @@ class IdeaController extends Controller
     {
 
 
-      $validated =  request()->validate([
+        $validated = request()->validate([
             'content' => 'required|min:5|max:240'
         ]);
 
 
-     Idea::create($validated);
+        $validated['user_id'] = auth()->id();
+
+        Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea was created successfully!');
 
@@ -34,6 +36,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', 'Idea was Deleted successfully!');
@@ -42,13 +48,23 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
+
         $editing = true;
 
         return view('ideas.show', compact('idea', 'editing'));
     }
     public function update(Idea $idea)
     {
-        $validated =  request()->validate([
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
+
+        $validated = request()->validate([
             'content' => 'required|min:5|max:240'
         ]);
 
@@ -56,7 +72,7 @@ class IdeaController extends Controller
         $idea->update($validated);
 
 
-        return redirect()->route('ideas.show' , $idea->id)->with('success', 'Idea was Updated successfully!');
+        return redirect()->route('ideas.show', $idea->id)->with('success', 'Idea was Updated successfully!');
 
     }
